requirements.txt
# Core dependencies
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1

# Deep learning frameworks
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
tensorflow-gpu==2.13.0

# Web framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# Image processing
scikit-image==0.21.0
scipy==1.11.1

# Machine learning
scikit-learn==1.3.0

# Configuration
pyyaml==6.0.1
configparser==5.3.0

# Utilities
requests==2.31.0
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Jetson specific
jetson-stats==4.0.0
jetson-utils==1.0.0

# Additional dependencies for Jetson
gstreamer-python==1.0.0
